worker_node/runpod_worker/requirements.txt
boto3==1.28.57
botocore==1.31.57
faster-whisper==0.6.0
PyYAML==6.0

